# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "DataPorts/Data Governance",
        "description": "DataPorts/Data Governance component",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "",
    "basePath": "",
    "schemes": [
        "http", "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/queryLoggerMetadata": {
            "get": {
                "tags": [
                    "Dataset metadata history"
                ],
                "summary": "Query the Data Governance dataset metadata history",
                "description": "List all the Data Governance dataset metadata history",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name" :"Authorization",
                        "in" : "header",
                        "type": "string",
                        "description" : "My token",
                        "required" : "true"
       
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryLoggerAllDatasourceMetadata": {
            "get": {
                "tags": [
                    "Dataset metadata history"
                ],
                "summary": "Query the Data Source dataset metadata history",
                "description": "List all the Data Source dataset metadata history",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name" :"Authorization",
                        "in" : "header",
                        "type": "string",
                        "description" : "My token",
                        "required" : "true"
       
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsProvider": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the user level access requests history per provider",
                "description": "List all the user level access requests history per provider",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsProviderByDataset/{datasetName}": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the user level access requests history per provider and by dataset name",
                "description": "List all the user level access requests history per provider by dataset name",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                },
                {
                    "in": "path",
                    "required": true,
                    "name": "datasetName",
                    "description": "Dataset metadata name",
                    "type": "string"
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsConsumer": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the user level access requests history per consumer",
                "description": "List all the user level access requests history per consumer",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsConsumerByDataset/{datasetName}": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the user level access requests history per consumer and by dataset name",
                "description": "List all the user level access requests history per consumer by dataset name",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                },
                {
                    "in": "path",
                    "required": true,
                    "name": "datasetName",
                    "description": "Dataset metadata name",
                    "type": "string"
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsOrgProvider": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the org level access requests history per provider",
                "description": "List all org user level access requests history per provider",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsOrgProviderByDataset/{datasetName}": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the org level access requests history per provider and by dataset name",
                "description": "List all org user level access requests history per provider by dataset name",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                },
                {
                    "in": "path",
                    "required": true,
                    "name": "datasetName",
                    "description": "Dataset metadata name",
                    "type": "string"
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/queryAccessRequestsOrgConsumer": {
            "get": {
                "tags": [
                    "Access requests history user & org level"
                ],
                "summary": "Get all the org level access requests history per consumer",
                "description": "List all the org level access requests history per consumer",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
...
    },
...
}